restructure our project structure
In order to run unit tests, we need to be able to modify the given build.gradle. This commit does that by creating a copy of the TeamCode
directory inside of the Automotons24-25 directory. Future younguns shouldn't pull the TeamCode directory from FTC and instead just use our Automotonus24-25 directory.